Move maxConnections into RDSClientConfig
