+++ pipeline/id_minter/src/test/scala/weco/pipeline/id_minter/fixtures/IdentifiersDatabase.scala
@@ -83,7 +83,8 @@ trait IdentifiersDatabase
     replicaHost = rdsHost,
     port = rdsPort,
     username = rdsUsername,
-    password = rdsPassword
+    password = rdsPassword,
+    maxConnections = rdsMaxPoolSize
   )
 
   def withIdentifiersTable[R](
